==> notebooks/09-19-parse-xml-create-article-text-dataframe.ipynb <==
## SemEval 2019 Task 4 - Parse XML and load article text into dataframe

Jonathan Miller and Negar Adyaniyazdi, VCU, CMSC516, Fall 2018

import xml.etree.cElementTree as et
import pandas as pd

DATA_PATH = '../data/'
DATA_RAW_PATH = DATA_PATH + 'raw/'

text_train_path = DATA_RAW_PATH + 'articles-training-20180831.xml'
text_val_path = DATA_RAW_PATH + 'articles-validation-20180831.xml'

Parse XML file for validation article data and create dataframe with ID and raw article text

tree = et.parse(text_val_path)
root = tree.getroot()
articles = root.findall('.//article')

article_text_list = []

for article in articles:
    text_string = article.get('title') + ' '
    for t in article.itertext():
        text_string += t
    article_text_list.append(text_string)

xml_data = [[article.get('id')] for article in articles]
col_names = ['id']
article_text_val = pd.DataFrame(xml_data, columns=col_names)
article_text_val['article_text'] = article_text_list
article_text_val['id'] = article_text_val['id'].astype(str)

article_text_val.head()

article_text_val.shape

Parse XML file for training article data and create dataframe with ID and raw article text

tree = et.parse(text_train_path)
root = tree.getroot()
articles = root.findall('.//article')

article_text_list = []

for article in articles:
    text_string = article.get('title') + ' '
    for t in article.itertext():
        text_string += t
    article_text_list.append(text_string)
    
xml_data = [[article.get('id')] for article in articles]
col_names = ['id']
article_text_train = pd.DataFrame(xml_data, columns=col_names)
article_text_train['article_text'] = article_text_list
article_text_train['id'] = article_text_train['id'].astype(str)

article_text_train.head()

article_text_train.shape

DATA_INTERIM_PATH = DATA_PATH + 'interim/'

article_text_train.to_csv(DATA_INTERIM_PATH + 'text_train.csv', index=False)
article_text_val.to_csv(DATA_INTERIM_PATH + 'text_val.csv', index=False)
