==> notebooks/09-19-parse-xml-create-metadata.ipynb <==
## SemEval 2019 Task 4 - Parse XML and load article metadata into dataframe

Jonathan Miller and Negar Adyaniyazdi, VCU, CMSC516, Fall 2018

import xml.etree.cElementTree as et
import pandas as pd

Get file paths and column names for ground truth and article text XML files

DATA_PATH = '../data/'
DATA_RAW_PATH = DATA_PATH + 'raw/'

gt_train_path = DATA_RAW_PATH + 'ground-truth-training-20180831.xml'
gt_val_path = DATA_RAW_PATH + 'ground-truth-validation-20180831.xml'
gt_cols = ['id', 'hyperpartisan', 'bias', 'url', 'labeled-by']

text_train_path = DATA_RAW_PATH + 'articles-training-20180831.xml'
text_val_path = DATA_RAW_PATH + 'articles-validation-20180831.xml'
text_cols = ['id', 'published-at', 'title']

Parse ground truth XML files

tree = et.parse(gt_train_path)
root = tree.getroot()
articles = root.findall('.//article')

xml_data = [[article.get('id'), article.get('hyperpartisan'), article.get('bias'), article.get('url'), article.get('labeled-by')] 
            for article in articles]

gt_train = pd.DataFrame(xml_data, columns=gt_cols)
gt_train['id'] = gt_train['id'].astype(str)

gt_train.head()

gt_train.shape

tree = et.parse(gt_val_path)
root = tree.getroot()
articles = root.findall('.//article')

xml_data = [[article.get('id'), article.get('hyperpartisan'), article.get('bias'), article.get('url'), article.get('labeled-by')] 
            for article in articles]

gt_val = pd.DataFrame(xml_data, columns=gt_cols)
gt_val['id'] = gt_val['id'].astype(str)

gt_val.head()

gt_val.shape

Parse Article XML files

tree = et.parse(text_train_path)
root = tree.getroot()
articles = root.findall('.//article')

xml_data = [[article.get('id'), article.get('published-at'), article.get('title')] 
            for article in articles]

text_train = pd.DataFrame(xml_data, columns=text_cols)
text_train['id'] = text_train['id'].astype(str)

text_train.head()

text_train.shape

tree = et.parse(text_val_path)
root = tree.getroot()
articles = root.findall('.//article')

xml_data = [[article.get('id'), article.get('published-at'), article.get('title')] 
            for article in articles]

text_val = pd.DataFrame(xml_data, columns=text_cols)
text_val['id'] = text_val['id'].astype(str)

text_val.head()

text_val.shape

Merge article and ground truth dataframes on index

md_train = text_train.merge(gt_train, on='id')
md_val = text_val.merge(gt_val, on='id')

Examine test set dataframe and columns

md_train.info()

md_train['hyperpartisan'].value_counts()

md_train['bias'].value_counts()

md_train['labeled-by'].value_counts()

Write metadata dataframes as csv

DATA_INTERIM_PATH = DATA_PATH + 'interim/'

md_train.to_csv(DATA_INTERIM_PATH + 'md_train.csv', index=False)
md_val.to_csv(DATA_INTERIM_PATH + 'md_val.csv', index=False)

md_train.head()
