==> notebooks/09-23-combine-data.ipynb <==
## SemEval 2019 Task 4 - Combine data files for convenience

Jonathan Miller and Negar Adyaniyazdi, VCU, CMSC516, Fall 2018

Load data

import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'

text_train_path = DATA_INTERIM_PATH + 'text_train.csv'
text_val_path = DATA_INTERIM_PATH + 'text_val.csv'

md_train_path = DATA_INTERIM_PATH + 'md_train.csv'
md_val_path = DATA_INTERIM_PATH + 'md_val.csv'

text_train = pd.read_csv(DATA_INTERIM_PATH + 'text_train.csv', dtype={'id' : str})
text_val = pd.read_csv(DATA_INTERIM_PATH + 'text_val.csv', dtype={'id' : str})

md_train = pd.read_csv(DATA_INTERIM_PATH + 'md_train.csv', dtype={'id' : str})
md_val = pd.read_csv(DATA_INTERIM_PATH + 'md_val.csv', dtype={'id' : str})

train = md_train.merge(text_train, on='id')
val = md_val.merge(text_val, on='id')

train.head()

train.shape

val.head()

val.shape

train.to_csv(DATA_INTERIM_PATH + 'train.csv', index=False)
val.to_csv(DATA_INTERIM_PATH + 'val.csv', index=False)
