==> notebooks/09-23-create-sample-dataframes.ipynb <==
# SemEval 2019 Task 4 - Generate Training Data Samples

Jonathan Miller and Negar Adyaniyazdi, VCU, CMSC516, Fall 2018

import pandas as pd
import numpy as np

import sys
sys.path.append('../src/data')

%load_ext autoreload
%autoreload 1

import preprocess
%aimport preprocess

Load data

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train.csv', dtype={'id' : str})
val = pd.read_csv(DATA_INTERIM_PATH + 'val.csv', dtype={'id' : str})

Examine 'bias' distribution of original data

train['bias'].value_counts(normalize=True)

val['bias'].value_counts(normalize=True)

Generate and preprocess some samples of data

N_SAMPLES_TRAIN = 2
SAMPLE_SIZE = 100
TRAIN_OR_VAL = 'val'
df_samples = []

for i in range(N_SAMPLES_TRAIN):
    if TRAIN_OR_VAL == 'train':
        df = train
    elif TRAIN_OR_VAL == 'val':
        df = val
    df_sample = df.sample(n=SAMPLE_SIZE, random_state=i)
    df_sample['preprocessed_text'] = preprocess.normalize_corpus(df_sample['article_text'])
    df_samples.append(df_sample)
    
df_samples[0].head()

DATA_PROCESSED_PATH = DATA_PATH + 'processed/'

for i in range(len(df_samples)):
    df_samples[i].to_csv(DATA_PROCESSED_PATH + TRAIN_OR_VAL + str(SAMPLE_SIZE) + '_' + str(i) + '.csv', index=False)

train100_0 = pd.read_csv(DATA_PROCESSED_PATH + 'train100_0.csv', dtype={'id' : str})

train100_0.head()

train100_0.loc[0,'article_text']

train100_0.loc[0,'preprocessed_text']

train[train['url'].str.contains('foxbusiness')].head()

gps = train.groupby('bias')

gps.groups['least']

from math import floor

sample_df = pd.DataFrame()
bias = train['bias'].value_counts(normalize=True)
gps = train.groupby('bias')
n = 10000
for group in gps.groups.keys():
    nrows = floor(n*bias[group])
    sample = train.loc[gps.groups[group],:].sample(n=nrows, random_state=1)
    sample_df = sample_df.append(sample)

sample_df['bias'].value_counts()

gps.groups['least']
