==> notebooks/09-26-find-nonenglish.ipynb <==
Load data and preprocessing module

import pandas as pd

import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from datatasks.custom_features import detect_language
%aimport datatasks.custom_features

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val.csv')

train['language'] = train['article_text'].apply(detect_language)

train['language'].value_counts()

non_english = train[~(train['language'] == 'english')]
non_english.reset_index(inplace=True, drop=True)
non_english.to_csv(DATA_PATH + 'processed/' + 'nonenglish_train.csv', index=False)

train.to_csv(DATA_PATH + 'interim/' + 'train.csv', index=False)

non_english['language'].value_counts()

non_english['hyperpartisan'].value_counts()

non_english[(non_english['hyperpartisan'] == True) & (non_english['title'].str.contains('National Liberty Federation') == False)]

val['language'] = val['article_text'].apply(detect_language)

non_english = val[~(val['language'] == 'english')]
non_english.reset_index(inplace=True, drop=True)
non_english.to_csv(DATA_PATH + 'processed/' + 'nonenglish_val.csv', index=False)

val.to_csv(DATA_PATH + 'interim/' + 'val.csv', index=False)

train.drop('published-at', axis=1)

train.head()
