==> notebooks/09-26-headline-sentiment-analysis.ipynb <==
Load data

import pandas as pd
import numpy as np

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train.csv')

from textblob import TextBlob

sample = train.sample(10000, random_state=1)
sample = sample[~(sample['title'].isna())]
sample.reset_index(inplace=True)

sample.info()

sample_title = sample.iloc[0]['title']
sample_title

blob = TextBlob(sample_title)

blob.sentiment

sample['polarity'] = sample['title'].apply(lambda x: TextBlob(x).sentiment[0])

sample['abs_polarity'] = np.absolute(sample['polarity'])

np.mean(sample['polarity'])

np.mean(sample['abs_polarity'])

np.mean(sample[sample['hyperpartisan'] == True]['abs_polarity'])

np.mean(sample[sample['hyperpartisan'] == False]['abs_polarity'])

sample['subjectivity'] = sample['title'].apply(lambda x: TextBlob(x).sentiment[1])

np.mean(sample['subjectivity'])

np.mean(sample[sample['hyperpartisan'] == True]['subjectivity'])

np.mean(sample[sample['hyperpartisan'] == False]['subjectivity'])

sample['polarity_t'] = sample['article_text'].apply(lambda x: TextBlob(x).sentiment[0])

sample['abs_polarity_t'] = np.absolute(sample['polarity_t'])

np.mean(sample['polarity_t'])

np.mean(sample['abs_polarity_t'])

np.mean(sample[sample['hyperpartisan'] == True]['abs_polarity_t'])

np.mean(sample[sample['hyperpartisan'] == False]['abs_polarity_t'])
