==> notebooks/09-27-explore-nonenglish.ipynb <==
import pandas as pd
import numpy as np

DATA_PATH = '../data/'
nonenglish = pd.read_csv(DATA_PATH + 'processed/' + 'nonenglish_train.csv')
nonenglishval = pd.read_csv(DATA_PATH + 'processed/' + 'nonenglish_val.csv')

nonenglish.head()

nonenglish['hyperpartisan'].value_counts()

pd.crosstab(nonenglish['hyperpartisan'], nonenglish['language'])

nonenglish[nonenglish['language'] == 'arabic']['article_text'][5930]

nonenglishval.head()

pd.crosstab(nonenglishval['hyperpartisan'], nonenglishval['language'])

nonenglish['id'].to_csv(DATA_PATH + 'processed/' + 'nonenglish_id_train.csv')
nonenglishval['id'].to_csv(DATA_PATH + 'processed/' + 'nonenglish_id_val.csv')

nonenglish.shape

nonenglishval.shape
