==> notebooks/09-27-generate-new-samples.ipynb <==
import pandas as pd

import sys
sys.path.append('../src/data')

%load_ext autoreload
%autoreload 1

import sample_data
%aimport sample_data

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val.csv')

train.head()

train = train[train['language'] == 'english']
val = val[val['language'] == 'english']

train.shape

val.shape

samples = sample_data.sample_preprocess_data(train, 2, 10000, 'train', save=True)

train.head()

train[train['url'].str.contains('beast') & (train['hyperpartisan'] == True)]

train[train['article_text'].str.contains('parry')]['hyperpartisan'].value_counts()
