==> notebooks/10-12-change-parsing-method.ipynb <==
import xml.etree.cElementTree as et
import pandas as pd

DATA_PATH = '../data/'
DATA_RAW_PATH = DATA_PATH + 'raw/'

text_train_path = DATA_RAW_PATH + 'articles-training-20180831.xml'
text_val_path = DATA_RAW_PATH + 'articles-validation-20180831.xml'

tree = et.parse(text_val_path)
root = tree.getroot()
articles = root.findall('.//article')

article = articles[0]

p_list = article.findall('p')

for p in p_list:
    links = p.findall('a')
    for a in links:
        link = a.get('href')
        print(a.get('href'), a.text)

external_links_lists = []
internal_links_lists = []

for article in articles[0:10]:
    
    external_links_dict = {}
    internal_links_dict = {}

    p_list = article.findall('p')

    for p in p_list:
        links = p.findall('a')
        for a in links:
            link = a.get('href')
            text = a.text
            if a.get('type') == 'external' and bool(link):
                if link[0] == '/':
                    internal_links_dict[link] = text
                else:
                    external_links_dict[link] = text
                    
    external_links_lists.append(external_links_dict)
    internal_links_lists.append(internal_links_dict)

internal_links_lists

text_string = article.get('title') + ' '
for t in article.itertext():
    text_string += t
    
text_string

DATA_INTERIM_PATH = DATA_PATH + 'interim/'
DATA_PROCESSED_PATH = DATA_PATH + 'processed/'

val = pd.read_csv(DATA_INTERIM_PATH + 'val.csv')
val_s = pd.read_csv(DATA_PROCESSED_PATH + 'val2500_0.csv')

val_s.iloc[0]

val[val['article_text'].str.contains('Advice for Hezbollah')]

val.iloc[3]['title']

val[val['article_text'].str.contains('FUCK')]['hyperpartisan'].value_counts()
