==> notebooks/10-12-explore-links.ipynb <==
import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val.csv')

train.head()

Average external links

import ast
len(ast.literal_eval(train['external_links'][0]))

e_links_count_HP = 0
for links in train[train['hyperpartisan']==True]['external_links']:
    e_links_count_HP += len(ast.literal_eval(links))
    
e_links_count_HP/train[train['hyperpartisan']==True].shape[0]

e_links_count_nonHP = 0
for links in train[train['hyperpartisan']==False]['external_links']:
    e_links_count_nonHP += len(ast.literal_eval(links))
    
e_links_count_nonHP/train[train['hyperpartisan']==False].shape[0]

i_links_count_HP = 0
for links in train[train['hyperpartisan']==True]['internal_links']:
    i_links_count_HP += len(ast.literal_eval(links))
    
i_links_count_HP/train[train['hyperpartisan']==True].shape[0]

i_links_count_nonHP = 0
for links in train[train['hyperpartisan']==False]['internal_links']:
    i_links_count_nonHP += len(ast.literal_eval(links))
    
i_links_count_nonHP/train[train['hyperpartisan']==False].shape[0]

HP vs. nonHP

TRAIN:

3.37 vs. 2.09 external links on average

0.17 vs. 0.12 internal links on average

VAL:

4.80 vs. 6.55 external links on average

e_links_count_HP = 0
for links in val[val['hyperpartisan']==True]['external_links']:
    e_links_count_HP += len(ast.literal_eval(links))
    
e_links_count_HP/val[val['hyperpartisan']==True].shape[0]

e_links_count_nonHP = 0
for links in val[val['hyperpartisan']==False]['external_links']:
    e_links_count_nonHP += len(ast.literal_eval(links))
    
e_links_count_nonHP/val[val['hyperpartisan']==False].shape[0]

i_links_count_HP = 0
for links in val[val['hyperpartisan']==True]['internal_links']:
    i_links_count_HP += len(ast.literal_eval(links))
    
i_links_count_HP/val[val['hyperpartisan']==True].shape[0]

i_links_count_nonHP = 0
for links in val[val['hyperpartisan']==False]['internal_links']:
    i_links_count_nonHP += len(ast.literal_eval(links))
    
i_links_count_nonHP/val[val['hyperpartisan']==False].shape[0]

train['num_external_links'] = train.loc[:,'external_links'].apply(lambda x: len(ast.literal_eval(x)))

train[['external_links', 'num_external_links']].head()

DATA_PROCESSED_PATH = DATA_PATH + 'processed/'
train = pd.read_csv(DATA_PROCESSED_PATH + 'train.csv')
val = pd.read_csv(DATA_PROCESSED_PATH + 'val.csv')

### Extracting domains

import tldextract

tldextract.extract('https://www.nbcnews.com/news/')

train['domain'] = train.loc[:,'url'].apply(lambda x: tldextract.extract(x)[1])

train.head()

train['domain'].value_counts()
