==> notebooks/10-13-domain-signatures.ipynb <==
### Domain 'signatures'

Attempting to find per-publisher 'signatures', or phrases that appear in most-all of a particular publisher's articles. Models are likely heavily overfitting on these terms.

import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val_c.csv')

train.head()

train['domain'].value_counts()

train[train['domain'] == 'foxbusiness'].reset_index()['article_text'][0]

train[(train['domain'] == 'foxbusiness') & (train['article_text'].str.contains('Continue Reading Below'))].shape[0] / train[train['domain'] == 'foxbusiness'].shape[0]

#### Every single Fox Business article in the dataset contains the phrase 'Continue Reading Below'

df = train[train['domain'] == 'foxbusiness'].reset_index()[0:500]

df.shape

import numpy as np
from sklearn.feature_extraction.text import CountVectorizer

vectorizer = CountVectorizer(stop_words='english',binary=True)
X = vectorizer.fit_transform(df['article_text'])

X = X.toarray()

sums = X.sum(axis=0)

sums.max()

thresh = 0.8
max_indices = np.argwhere(sums >= thresh*df.shape[0])
print(max_indices)

for index in max_indices:
    feature = vectorizer.get_feature_names()[index[0]]
    print(feature)

df[df['article_text'].str.contains('below')].reset_index()['article_text'][2]

import numpy as np
from sklearn.feature_extraction.text import CountVectorizer

# Extract and inspect words that appear in some (large) percentage of a publisher's articles
def find_domain_signatures(df, domain, thresh, stop_words='english', should_print=True):
    
    # Get all articles by provided publisher
    df = df[df['domain'] == domain]
    
    # Create counter vectorizer
    vectorizer = CountVectorizer(stop_words=stop_words,binary=True)
    X = vectorizer.fit_transform(df['article_text'])
    
    # Convert to dense array
    X = X.toarray()
    
    # Create sums vector with entry for each word
    sums = X.sum(axis=0)
    
    # Find all indices which exceed the given percentage for inclusion
    thresh = thresh
    max_indices = np.argwhere(sums >= thresh*df.shape[0])
    
    signature_tokens = []
    
    # Find words at indices and append to list
    for index in max_indices:
        feature = vectorizer.get_feature_names()[index[0]]
        signature_tokens.append(feature)
    
    # Print and return
    if should_print:
        print(signature_tokens)
    return signature_tokens

import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from models.EDA import find_domain_signatures
%aimport models.EDA

sig = find_domain_signatures(train, 'foxbusiness', thresh=0.5)

train[(train['domain']=='thedailybeast') & (train['article_text'].str.contains('beast'))].reset_index()['url'][0]
