==> notebooks/10-13-explore-caps.ipynb <==
import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val_c.csv')

import re
caps = r"\b[A-Z][A-Z]+\b"

train['num_caps_words'] = train.loc[:,'article_text'].apply(lambda x: len(re.findall(caps,x)))

t = train[train['hyperpartisan']==True].reset_index()['article_text']

t[0]

def compare_text(text, text_type='article_text'):
    print('Train HP: ' + str(train[(train['article_text'].str.contains(text)) & (train['hyperpartisan']==True)].shape[0]))
    print('Train nonHP: ' + str(train[(train['article_text'].str.contains(text)) & (train['hyperpartisan']==False)].shape[0]))
    print('Val HP: ' + str(val[(val['article_text'].str.contains(text)) & (val['hyperpartisan']==True)].shape[0]))
    print('Val nonHP: ' + str(val[(val['article_text'].str.contains(text)) & (val['hyperpartisan']==False)].shape[0]))

compare_text('a lot')

train[(train['article_text'].str.contains('BITCH')) & (train['hyperpartisan']==True)].reset_index()['url'][3]

import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from models.EDA import compare_text
%aimport models.EDA

compare_text('Fox News', train, val)

compare_text('memes', train, val)
