==> notebooks/10-15-consecutive-caps.ipynb <==
import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')

import re
regex = r'[A-Z]+(?=\s[A-Z])(?:\s[A-Z]+)+'

re.findall(regex, 'THIS IS A string I am USING TO TEST DUDE')

def consecutive_caps_words(text):
    regex = r'[A-Z]+(?=\s[A-Z])(?:\s[A-Z]+)+'
    matches = re.findall(regex, text)
    n_consec = 0
    for match in matches:
        tokens = match.split()
        n_tokens = len(tokens)
        if n_tokens > n_consec:
            n_consec = n_tokens
            
    return n_consec

train['consecutive_caps'] = train['article_text'].apply(consecutive_caps_words)

train['consecutive_caps'].head()

import numpy as np

train[train['hyperpartisan']==True]['consecutive_caps'].mean()

train[train['hyperpartisan']==False]['consecutive_caps'].mean()

train[(train['hyperpartisan']==True) & (train['consecutive_caps'] > 10)].shape

train[(train['hyperpartisan']==False) & ~(train['domain']=='reuters') & (train['consecutive_caps'] > 10)].shape

train[(train['hyperpartisan']==False) & ~(train['domain']=='reuters') & (train['consecutive_caps'] > 10)].reset_index()['article_text'][0]
