==> notebooks/10-15-identify-signatures.ipynb <==
import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from models.EDA import find_domain_signatures
%aimport models.EDA

import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')

domains = train['domain'].value_counts().keys()
domains

Initialize empty domain keywords list and iterate

signatures = {}
for domain in domains:
    sig = find_domain_signatures(train, domain, should_print=False)
    signatures[domain] = sig

signatures['dcclothesline']

from sklearn.feature_extraction.text import CountVectorizer

vectorizer = CountVectorizer(stop_words='english',binary=True)
X = vectorizer.fit_transform(train['article_text'])

features = vectorizer.get_feature_names()
features.index('according')

X[:,701867].sum()/train.shape[0]

domain = 'truthdig'
features = vectorizer.get_feature_names()
diffs = {}
for key in signatures[domain].keys():
    index = features.index(key)
    percent = X[:,index].sum()/train.shape[0]
    diff = round(float(signatures[domain][key]) - percent, 3)
    diffs[key] = diff

from models.EDA import filter_common_signatures
%aimport models.EDA

diffs = filter_common_signatures(signatures, vectorizer, X)

for key in list(diffs.keys())[0:30]:
    
    print(key, diffs[key])

domain = 'foxbusiness'
df = train[train['domain']==domain]
df.reset_index(inplace=True)
df['article_text'] = df['article_text'].str.lower()

word = 'opens a new window'
df[(df['article_text'].str.contains(word))].reset_index()['article_text'].replace(word, word.upper()).shape

df.shape

sample = df.sample(n=1000)
vectorizer = CountVectorizer(stop_words='english',binary=True,ngram_range=(2,2))
X = vectorizer.fit_transform(sample['article_text'])

sums = X.sum(axis=0)

feature_names = vectorizer.get_feature_names()

import numpy as np

thresh = 0.15
max_indices = np.argwhere(sums >= thresh*df.shape[0])

for index in max_indices:
    feature = feature_names[index[1]]
    print(feature)
    print(round(sums[0,index[1]]/df.shape[0], 3))

feature_names[2149224]

sums.shape

domain = 'thedailybeast'
uni_sigs = find_domain_signatures(train, domain)
bi_sigs = find_domain_signatures(train, domain, thresh=0.15, ngram_range=(2,2))

train[(train['domain']=='thedailybeast') & (train['article_text'].str.contains('speedy'))].reset_index()['article_text'][0]

train[train['article_text'].str.lower().str.contains(' lta ')]['domain'].value_counts()

train[train['domain']=='foxbusiness' & (train['article_text'].str.lower().str.contains(' lta '))].reset_index()['article_text'][0]
