==> notebooks/10-15-scrape-mediabias.ipynb <==
import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from externals.bias_list import get_bias_domains, get_bias_tlds
%aimport externals.bias_list

domains_dict = get_bias_domains()

tlds = get_bias_tlds(domains_dict)

import pickle

path = '../data/external/'

pickle_out = open(path + "dict.pickle","wb")
pickle.dump(tlds, pickle_out)
pickle_out.close()

pickle_in = open(path + "dict.pickle","rb")
example_dict = pickle.load(pickle_in)
