==> notebooks/10-16-linked-domains-bias.ipynb <==
import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')

import pickle

DATA_EXTERNAL_PATH = DATA_PATH + 'external/'
pickle_in = open(DATA_EXTERNAL_PATH + "tld.pickle","rb")
tlds = pickle.load(pickle_in)

import ast

ast.literal_eval(train['external_links'][0])

from tldextract import extract

def external_links_bias(external_links, tlds) :
    
    HP_links = 0
    nonHP_links = 0
    unknown_links = 0

    for url in ast.literal_eval(external_links):
        tld = extract(url)[1]
        try:
            bias = tlds[tld]
            if bias in ['left', 'right']:
                HP_links += 1
            elif bias in ['leftcenter', 'right-center', 'center']:
                nonHP_links += 1
        except:
            unknown_links += 1
            
    return pd.Series([HP_links, nonHP_links, unknown_links])

train[['HP_links_count', 'nonHP_links_count', 'unknown_links_count']] = train.loc[:,'external_links'].apply(external_links_bias, )

for HP in [True, False]:
    df = train[train['hyperpartisan'] == HP]
    df_shape = df.shape[0]
    HP_links = df['HP_links_count'].sum()
    nonHP_links = df['nonHP_links_count'].sum() 
    unknown_links = df['unknown_links_count'].sum()
    avg_links = (HP_links + nonHP_links + unknown_links)/df_shape
    avg_HP_links = HP_links / df_shape
    avg_nonHP_links = nonHP_links / df_shape
    print(HP)
    print('Average external links = {}'.format(avg_links))
    print('Average HP links = {}'.format(avg_HP_links))
    print('Average nonHP links = {}'.format(avg_nonHP_links))

train[train['hyperpartisan']==False]['HP_links_count'].value_counts()

parry = train[train['article_text'].str.lower().str.contains('fusion')].reset_index()

parry['domain'].value_counts()

parry[parry['domain']=='newsline'].reset_index()['article_text'][0]
