==> notebooks/10-17-signatures.ipynb <==
import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')

train.loc[:,'article_text_lower'] = train['article_text'].str.lower()

domains = train['domain'].value_counts().keys()
domains

import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from models.EDA import find_domain_signatures
%aimport models.EDA

def sigs(domain):
    uni_sigs = find_domain_signatures(train, domain)
    bi_sigs = find_domain_signatures(train, domain, thresh=0.15, ngram_range=(2,2))

current_domain = domains[15]
print(current_domain)
sigs(current_domain)

article_text = train[(train['domain']==current_domain) & (train['article_text_lower'].str.contains('michael key'))].reset_index()['article_text']

article_text[0]

### General:
urls, email addresses, twitter usernames

### foxbusiness:

'Continue Reading Below', 
'Advertisement', 
'Image source:',
'Click here',
'Opens a New Window..'
'disclosure policy'
'The Motley Fool'
'has no position',
'stocks mentioned'

### abqjournal
'ALBEQUERQE, N.M.'
'ADVERTISEMENT'
'new mexico'

### apnews
'(AP)'

### newsline
'Fusion Media or anyone involved with Fusion Media will not accept any liability for loss or damage as a result of reliance on the information including data, quotes, charts and buy/sell signals contained within this website. Please be fully informed regarding the risks and costs associated with trading the financial markets, it is one of the riskiest investment forms possible.'

### motherjones
first line is often broken link like '&lt;a href="http://www.flickr.com/photos/talkradionews/2510926965/sizes/l/in/photostream/"&gt;TalkRadioNews&lt;/a&gt;/Flickr'

### thedailybeast
'Start and finish your day with the top stories from The Daily Beast. A speedy, smart summary of all the news you need to know (and nothing you don't).'

### nbcnews
'nbc'

### reuters
come back to this

### consortiumnews
'Robert Parry'
