==> notebooks/10-23-incorporate-custom-features.ipynb <==
import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from models.models import create_tfidf, create_custom_features, run_models
%aimport models.models

import pandas as pd

DATA_PATH = '../data/'
DATA_PROCESSED_PATH = DATA_PATH + 'processed/'
train = pd.read_csv(DATA_PROCESSED_PATH + 'train100_0.csv')
val = pd.read_csv(DATA_PROCESSED_PATH + 'val25_0.csv')

X_train_tfidf, X_test_tfidf, y_train, y_test, tfidf_vectorizer = create_tfidf(train, val)

train.head()

X_custom, y_custom = create_custom_features(train, val, names=['HP_links_count', 'nonHP_links_count'])

run_models(['lr'], X_train_tfidf, X_test_tfidf, y_train, y_test, X_custom, y_custom)

import numpy as np
from scipy.sparse import hstack

hstack((X_train_tfidf, np.array(X_custom))).A
