==> notebooks/10-23-pipeline-refactor.ipynb <==
import pandas as pd

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'

import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from datatasks.sample_data import sample_data
%aimport datatasks.sample_data

from sklearn.pipeline import Pipeline

import glob
DATA_PROCESSED_PATH = DATA_PATH + 'processed/'

# Get training and test data
train_path = glob.glob(DATA_PROCESSED_PATH + 'train*.csv')[0]
val_path = glob.glob(DATA_PROCESSED_PATH + 'val*.csv')[0]

# Load training and validation data
train = pd.read_csv(train_path)
val = pd.read_csv(val_path)

train.head()

X_train = train.drop('hyperpartisan', axis=1)
y_train = train['hyperpartisan']
X_test = val.drop('hyperpartisan', axis=1)
y_test = val['hyperpartisan']

# CREDIT: https://www.kaggle.com/baghern/a-deep-dive-into-sklearn-pipelines

from sklearn.base import BaseEstimator, TransformerMixin

class TextSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Use on text columns in the data
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[self.key]
    
class NumberSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Use on numeric columns in the data
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[[self.key]]

from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import TfidfVectorizer

text = Pipeline([
    ('selector', TextSelector(key='preprocessed_text')),
    ('tfidf', TfidfVectorizer( stop_words='english'))
])

HP_links = Pipeline([
    ('selector', NumberSelector(key='HP_links_count'))
])

nonHP_links = Pipeline([
    ('selector', NumberSelector(key='nonHP_links_count'))
])

unknown_links = Pipeline([
    ('selector', NumberSelector(key='unknown_links_count'))
])

feats = FeatureUnion([
    ('text', text),
    ('HP_links', HP_links),
    ('nonHP_links', nonHP_links)
])

pipeline = Pipeline([
    ('features',feats),
    ('classifier', LogisticRegression(C=30.0, class_weight='None', solver='newton-cg')),
])

pipeline.fit(X_train, y_train)

preds = pipeline.predict(X_test)

from models.models import evaluate_model

evaluate_model(preds, y_test)

for df in [train, val]:
    for col in ['HP_links_count', 'nonHP_links_count', 'unknown_links_count']:
        df.loc[:,col] = df.loc[:,col].apply(lambda x: 1 if x > 0 else 0)

train[train['unknown_links_count']==1]['hyperpartisan'].value_counts()

type(X_train)
