==> notebooks/10-25-publisher-check.ipynb <==
import pandas as pd
DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_c.csv')

import pickle

UTIL_PATH = '../util/'
pickle_in = open(UTIL_PATH + "tld.pickle","rb")
tlds = pickle.load(pickle_in)

train_pubs = set(train['domain'].values)
all_pubs = set(tlds.keys())

list(train_pubs.difference(all_pubs))

train[train['domain'].isin(list(train_pubs.difference(all_pubs)))]['hyperpartisan'].value_counts()

train[(train['domain'].isin(list(train_pubs.difference(all_pubs)))) & (train['hyperpartisan']==False)]['domain'].value_counts()

train[(train['domain']).isin(list(train_pubs.difference(all_pubs))) & (train['hyperpartisan']==True)]['domain'].value_counts()

list(all_pubs.difference(train_pubs))
