==> notebooks/11-1-keras-tokenizer.ipynb <==
# Experiment with Keras Tokenizer

Seeing the efficiency of Keras tokenizer on tokenizing article text

#### Dependencies

import pandas as pd
from keras.preprocessing.text import Tokenizer

%load_ext autoreload
%autoreload 1

#### Read data

DATA_PATH = '../data/'
train = pd.read_csv(DATA_PATH + 'interim/train_c.csv')

t = Tokenizer()

docs = list(train['article_text'])

t.fit_on_texts(docs)

t.word_counts['libtard']

lib = train[train['article_text'].str.lower().str.contains('libtard')].reset_index()

lib[lib['hyperpartisan']==False]

dup = train.duplicated('url', keep=False)

train.drop_duplicates(subset='url', keep='first')

train.iloc[dup[dup==True].index,:]['url'][78]

train[train['url']=='http://aim.org/don-irvine-blog/cnn-balks-at-trump-campaign-ad-calling-media-enemies-video/']

dup[dup==True].shape[0]

767257 + 32743
