==> notebooks/11-2-test-instance-reduction.ipynb <==
import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from datatasks.remove_articles import remove_articles
%aimport datatasks.remove_articles

DATA_INTERIM_PATH = '../data/interim/'
remove_articles(DATA_INTERIM_PATH)

import pandas as pd

train = pd.read_csv(DATA_INTERIM_PATH + 'train_reduced.csv')

train.shape[0]

train['hyperpartisan'].value_counts(normalize=True)

pd.read_csv(DATA_INTERIM_PATH)

train.drop('Unnamed: 0', axis=1, inplace=True)
train.reset_index(inplace=True)
train.to_csv(DATA_INTERIM_PATH + 'train_reduced.csv')

train.drop('index', axis=1, inplace=True)

train.to_csv(DATA_INTERIM_PATH + 'train_reduced.csv', index=False)

train.head()
