==> notebooks/11-3-preprocessing.ipynb <==
# Preprocessing

#### Tasks:

- Add title to article text

- Remove links, HTML, twitter usernames, emails

- Remove publisher signatures ( TRAIN ONLY )

#### Dependencies

import pandas as pd
import numpy as np

import sys
sys.path.append('../src')

from datatasks.new_preprocess import preprocess

preprocess('../data/interim/')

DATA_INTERIM_PATH = '../data/interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train_p.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val_p.csv')

train.shape

val.shape

#### Add title to article text

def add_title_to_article_text(article):
    if article['title'] is not np.nan:
        return str(article['title']) + ' ' + str(article['article_text'])
    else:
        return str(article['article_text'])

for df in [train, val]:
    df['preprocessed_text'] = df.apply(add_title_to_article_text, axis=1)

#### Remove HTML artifacts

import re

def replace_html_stuff(article):
    preprocessed_text = article['preprocessed_text']
    
    # Remove all between left and right brackets
    preprocessed_text = re.sub(r'(?s)&lt.*?&gt;', '', preprocessed_text)
    preprocessed_text = re.sub(r'(?s)lt;.*?gt;', '', preprocessed_text)
    
    # Remove all html tags
    preprocessed_text = re.sub(r'&(.{1,5});', '', preprocessed_text)
    preprocessed_text = re.sub(r'amp;', '', preprocessed_text)
    
    # Remove !function
    preprocessed_text = re.sub(r'(?s)!function.*?(\?|\")\);', '', preprocessed_text)
    
    return preprocessed_text

for df in [train, val]:
    df['preprocessed_text'] = df.apply(replace_html_stuff, axis=1)

##### TODO: URLs, Twitter Usernames, Emails

train['domain'].value_counts().head(30)

from gensim.models.phrases import Phrases, Phraser
from nltk.tokenize.toktok import ToktokTokenizer

dailybeast = train[train['domain']=='circa'].reset_index()['preprocessed_text']

tokenizer = ToktokTokenizer()

all_tokens = [tokenizer.tokenize(dailybeast[i]) for i in range(len(dailybeast))]

phraser = Phraser(Phrases(all_tokens, min_count=len(dailybeast)/10, threshold=32))

phraser.phrasegrams

check = train[train['preprocessed_text'].str.contains('\[(.*)\]')]

check.reset_index()['preprocessed_text'][5]
