==> notebooks/11-6-publishers-in-val-not-train.ipynb <==
# Investigating publisher breakdown in train/val

Look specifically for publishers in val but not in train and add to evaluate model function to show how model performs on unseen publishers

import pandas as pd

DATA_INTERIM_PATH = '../data/interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train_p.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val_p.csv')

train_pubs = set(list(train['domain']))
val_pubs = set(list(val['domain']))

val_diff = val_pubs - train_pubs

val_diff

val[val['domain'].isin(val_diff)].shape

val['domain'].value_counts()

#### Looks like val is all new publishers...oops
