==> notebooks/11-11-examine-df-tlds-not-in-bias-list.ipynb <==
import pandas as pd
import pickle

DATA_INTERIM_PATH = 'data/interim/'

train = pd.read_csv(DATA_INTERIM_PATH + 'train_p.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val_p.csv')

util_path = 'util/'
pickle_in = open(util_path + "tld.pickle", "rb")
tlds = pickle.load(pickle_in)

train_domains = set(train['domain'])
val_domains = set(val['domain'])
df_domains = train_domains.union(val_domains)

len(df_domains)

scraped_domains = set(tlds.keys())

diff = df_domains.difference(scraped_domains)

diff_dict = {}
for domain in diff:
    if domain in train_domains:
        diff_dict[domain] = train[train['domain'] == domain]['bias'].values[0]
    elif domain in val_domains:
        diff_dict[domain] = val[val['domain'] == domain]['bias'].values[0]
        
    if diff_dict[domain] == 'left-center':
        diff_dict[domain] = 'leftcenter'
    elif diff_dict[domain] == 'least':
        diff_dict[domain] = 'center'
        
diff_dict

train[train['domain'] == 'foxbusiness']['bias'].values[0]

merged_dict = {**tlds, **diff_dict}

merged_dict['fury']

util_path = 'util/'

pickle_out = open(util_path + "tld.pickle", "wb")
pickle.dump(merged_dict, pickle_out)
pickle_out.close()

len(merged_dict.keys())

len(tlds.keys())
