==> notebooks/11-24-scattertext.ipynb <==
# Using scattertext to visualize data

#### Dependencies

import pandas as pd  
import spacy
import numpy as np
import scattertext as st

#### Load data

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
train = pd.read_csv(DATA_INTERIM_PATH + 'train_p.csv')
val = pd.read_csv(DATA_INTERIM_PATH + 'val_p.csv')

#### Create spaCy model

nlp = spacy.load('en_core_web_md')

#### Create Scattertext Corpus

corpus = st.CorpusFromPandas(train,
                            category_col='hyperpartisan',
                            text_col = 'preprocessed_text',
                            nlp=nlp).build()

help(st.produce_scattertext_explorer)
