==> notebooks/11-26-gensimify-project.ipynb <==
# Converting project to Gensim best pratices

#### Dependencies

import logging
import gensim
from gensim import corpora
from pprint import pprint
from gensim.models import TfidfModel
from gensim.utils import simple_preprocess
from gensim.sklearn_api import TfIdfTransformer
from smart_open import smart_open
import os
import pandas as pd
import numpy as np
import ast
from tqdm import tqdm
import pickle
import glob
import itertools
from gensim.sklearn_api import TfIdfTransformer
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression

import sys
sys.path.append('../src')

%load_ext autoreload
%autoreload 1

from models.models import evaluate_model

from models.plot import plot_LSA, plot_confusion_matrix

logging.basicConfig(format='%(levelname)s : %(message)s', level=logging.INFO)
logging.root.level = logging.INFO  # ipython sometimes messes up the logging setup; restore

#### Load data and get token lists from dataframe

This only needs to be run once. Token lists will be saved in order to take advantage of Gensim's loading

DATA_PATH = '../data/'
DATA_INTERIM_PATH = DATA_PATH + 'interim/'
DATA_PROCESSED_PATH = DATA_PATH + 'processed/'

#train = pd.read_csv(DATA_INTERIM_PATH + 'train_p.csv', usecols=['tokens'])
#train_tokens1 = [ast.literal_eval(train['tokens'].values[i]) for i in tqdm(range(0, 149999))]

#with open(DATA_PROCESSED_PATH + "train_tokens1.txt", "wb") as internal_filename:
#    pickle.dump(train_tokens1, internal_filename)

#val = pd.read_csv(DATA_INTERIM_PATH + 'val_p.csv', usecols=['tokens'])
#val_tokens = [ast.literal_eval(val['tokens'].values[i]) for i in tqdm(range(0,val.shape[0]))]
#with open(DATA_PROCESSED_PATH + "val_tokens.txt", "wb") as internal_filename:
#    pickle.dump(val_tokens, internal_filename)

#### Create Gensim Dictionary and Bag-of-Words Corpus

#train_dictionary = corpora.Dictionary()
#fnames = list(reversed(glob.glob(DATA_PROCESSED_PATH + 'train_tokens*')))
#for fname in fnames:
#    with open(fname, "rb") as internal_filename:
#        train_tokens = pickle.load(internal_filename)
#    train_dictionary.add_documents(train_tokens)

#train_dictionary.save(DATA_PROCESSED_PATH + 'train_dict.dict')

class MyCorpus(object):
    
    def __init__(self, filepath, dictionary, trainorval):
        self.filepath = filepath
        self.dictionary = dictionary
        self.trainorval = trainorval
        self.fnames = list(reversed(glob.glob(filepath + trainorval + '_tokens*')))
    
    def __iter__(self):
        for fname in self.fnames:
            with open(fname, "rb") as internal_filename:
                for doc in pickle.load(internal_filename):
                    yield self.dictionary.doc2bow(doc)

#### Serialize train and val BoW

train_dictionary = corpora.Dictionary.load(DATA_PROCESSED_PATH + 'train_dict.dict')
train_corpus = MyCorpus(DATA_PROCESSED_PATH, train_dictionary, 'train')
gensim.corpora.MmCorpus.serialize(DATA_PROCESSED_PATH + 'train_bow.mm', train_corpus)

val_corpus = MyCorpus(DATA_PROCESSED_PATH, train_dictionary, 'val')
gensim.corpora.MmCorpus.serialize(DATA_PROCESSED_PATH + 'val_bow.mm', train_corpus)

#### Load BoW

train_mm_corpus = gensim.corpora.MmCorpus(DATA_PROCESSED_PATH + 'train_bow.mm')
val_mm_corpus = gensim.corpora.MmCorpus(DATA_PROCESSED_PATH + 'val_bow.mm')

#### Load targets

y_train = pd.read_csv(DATA_INTERIM_PATH + 'train_p.csv', usecols=['hyperpartisan'])
y_test = pd.read_csv(DATA_INTERIM_PATH + 'val_p.csv', usecols=['hyperpartisan'])

#### TF-IDF

tfidf = TfIdfTransformer(dictionary=train_dictionary)
X_train = tfidf.fit_transform(train_mm_corpus)

#### Fit Classifier

clf = LogisticRegression()
clf.fit(X_train, y_train['hyperpartisan'].tolist())

#### Predict and Evaluate

X_test = tfidf.transform(val_mm_corpus)

preds = clf.predict(X_test, y_test['hyperpartisan'].tolist())

evaluate_model(y_test, preds)

plot_confusion_matrix(y_test, preds)
